thirteenth commit

+++ src/main/java/com/anjaniy/springbootcrud/models/Student.java
@@ -45,7 +45,7 @@ public class Student {
         return cgpa;
     }
 
-    public void setCgpa(int cgpa) {
+    public void setCgpa(float cgpa) {
         this.cgpa = cgpa;
     }
